+++ XrmPluginCore.Tests.Context/BusinessDomain/optionsets/account_address1_shippingmethodcode.cs
@@ -10,22 +10,27 @@ public enum account_address1_shippingmethodcode
 {
     [EnumMember]
     [OptionSetMetadata("Airborne", 1033)]
+    [OptionSetMetadata("Luftfragt", 1030)]
     Airborne = 1,
 
     [EnumMember]
     [OptionSetMetadata("DHL", 1033)]
+    [OptionSetMetadata("DHL", 1030)]
     DHL = 2,
 
     [EnumMember]
     [OptionSetMetadata("FedEx", 1033)]
+    [OptionSetMetadata("FedEx", 1030)]
     FedEx = 3,
 
     [EnumMember]
     [OptionSetMetadata("UPS", 1033)]
+    [OptionSetMetadata("UPS", 1030)]
     UPS = 4,
 
     [EnumMember]
     [OptionSetMetadata("Postal Mail", 1033)]
+    [OptionSetMetadata("Alm. post", 1030)]
     PostalMail = 5,
 
     [EnumMember]