+++ XrmPluginCore.Tests.Context/BusinessDomain/optionsets/contact_shippingmethodcode.cs
@@ -10,5 +10,6 @@ public enum contact_shippingmethodcode
 {
     [EnumMember]
     [OptionSetMetadata("Default Value", 1033)]
+    [OptionSetMetadata("Standardværdi", 1030)]
     DefaultValue = 1,
 }